version 1.1.0

+++ src/test/java/ContactDAOTest.java
@@ -2,6 +2,8 @@ import com.contactapp.dao.ContactDAO;
 import com.contactapp.model.Contact;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 public class ContactDAOTest {
@@ -11,36 +13,50 @@ public class ContactDAOTest {
     public void testAddContact() {
         Contact contact = new Contact(0, "Ana Cardona", "Ana123@example.com", "12345678", "1002609549");
         boolean result = dao.addContact(contact);
-        assertTrue(result, "El contacto se agrego correctamente");
+        System.out.println("Contacto agregado: " + contact);
+        assertTrue(result, "El contacto no se agregó correctamente.");
 
     }
 
     @Test
-    public void testGetAllContactsNoVacios() {
-
-        assertFalse(dao.getAllContacts().isEmpty(), "No hay contactos en la base de datos");
-
+    public void testGetAllContacts() {
+        List<Contact> contactos = dao.getAllContacts();
+        System.out.println(" Contactos encontrados: " + contactos);
+        assertTrue(contactos.size() > 0, "No se encontraron contactos en la base de datos.");
     }
+
 
     @Test
     public void testGetContactById() {
         int id = 2;
         Contact contact = dao.getContactById(id);
-        assertTrue(contact != null, "El contacto existe en la base de datos" + contact);
+        System.out.println("Contacto encontrado: " + contact);
+        assertFalse(contact == null, "El contacto no existe en la base de datos");
     }
 
     @Test
     public void testDeleteContact() {
-        int id = 6;
+        int id = 8;
+
+        Contact contact = dao.getContactById(id);
+
+        if (contact == null) {
+            throw new IllegalArgumentException(" El contacto con ID " + id + " no existe y no se puede eliminar.");
+        }
+
         boolean result = dao.deleteContact(id);
-        assertTrue(result, "El contacto se elimino correctamente");
+        System.out.println("Contacto eliminado: ID = " + id);
+        assertTrue(result, "Falló la eliminación del contacto con ID: " + id);
     }
+
 
     @Test
     public void testUpdateContact() {
         Contact contact = new Contact(3, "Pedro Picapiedra", "pedrito@example.com", "12345678", "1002609666");
+
         boolean result = dao.updateContact(contact);
-        assertTrue(result, "El contacto se actualizo correctamente");
+        System.out.println("Contacto actualizado: " + contact);
+        assertTrue(result, "El contacto no existe en la base de datos");
 
     }
 }
